+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/attributes/interactions/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains the various possible {@link io.github.thebusybiscuit.slimefun4.core.attributes.interactions.InteractionResult}s
+ * that can be returned by an {@link io.github.thebusybiscuit.slimefun4.core.attributes.ExternallyInteractable} object.
+ */
+package io.github.thebusybiscuit.slimefun4.core.attributes.interactions;
